fix bondExtra issue
and refactor
