+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/polkagate/snap.git"
   },
   "source": {
-    "shasum": "FA/uWVGqQw4/rWASk1AKXaPGIwfWSzoHGWHkByzdJtA=",
+    "shasum": "i+k81owzfZQiS9HxqaYowm6IoUUEhSaNBd4RmlFIJJI=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
