+++ packages/snap/src/ui/review/index.ts
@@ -1,6 +1,11 @@
 // Copyright 2023-2024 @polkagate/snap authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './txHeaderUseApi';
-export * from './txBody';
+export * from './metadataAlert';
 export * from './reviewUseApi';
+export * from './reviewUseMetadata';
+export * from './txBody';
+export * from './txContentUseApi';
+export * from './txContentUseMetadata';
+export * from './footer';
+export * from './header';
